changed DEBUG_HEAD & DEBUG_FOOT to preprocessor macros from const char*
this is more for when I copy and past this into skele_lib because that allows me to have default values without making it so that you have to go into my header to change the value.
this also allowed me to change GenerateDebugMessage(const char*, const char* const char*) to GenerateDebugMessage(const char*) making calling that much easier.